javadoc work in progress

+++ spartagold/wallet/backend/ProofOfWork.java
@@ -1,8 +1,16 @@
 package spartagold.wallet.backend;
 
-
 import java.io.UnsupportedEncodingException;
 import java.util.UUID;
+
+/**
+ * Performs the proof-of-work algorithm on a block string. Creates a random string, then concatenates
+ * random string to block string and an increasing long number. If a hash of this concatenation 
+ * contains a number of zeroes equal to NUMBER_OF_ZEROES, solution is found.
+ * 
+ * @author Art Tucay Jr., Paul Portela
+ * @version 1.0.0
+ */
 
 public class ProofOfWork extends Thread
 {
@@ -13,7 +21,7 @@ public class ProofOfWork extends Thread
 	{
 		String randomString = UUID.randomUUID().toString();
 
-		int i = 0;
+		long i = 0;
 		boolean isFound = false;
 		String zeroes = String.format(String.format("%%%ds", NUMBER_OF_ZEROES), " ").replace(" ", "0");
